Add README for the new APIs
